+++ src/lib/url-sync.js
@@ -95,6 +95,9 @@ class URLSync {
     this.mapping = options.mapping || {};
     this.getHistoryState = options.getHistoryState || (() => null);
     this.threshold = options.threshold || 700;
+    this.updateOnEveryKeyStroke = options.updateOnEveryKeyStroke !== undefined ?
+      options.updateOnEveryKeyStroke :
+      true;
     this.trackedParameters = options.trackedParameters || ['query', 'attribute:*', 'index', 'page', 'hitsPerPage'];
 
     this.searchParametersFromUrl = AlgoliaSearchHelper
@@ -108,7 +111,11 @@ class URLSync {
     // we need to create a REAL helper to then get its state. Because some parameters
     // like hierarchicalFacet.rootPath are then triggering a default refinement that would
     // be not present if it was not going trough the SearchParameters constructor
-    this.originalConfig = algoliasearchHelper({}, currentConfiguration.index, currentConfiguration).state;
+    this.originalConfig = algoliasearchHelper(
+      {addAlgoliaAgent() {}},
+      currentConfiguration.index,
+      currentConfiguration
+    ).state;
     return this.searchParametersFromUrl;
   }
 
@@ -121,6 +128,7 @@ class URLSync {
   }
 
   onPopState(helper, fullState) {
+    clearTimeout(this.urlUpdateTimeout);
     // compare with helper.state
     const partialHelperState = helper.getState(this.trackedParameters);
     const fullHelperState = assign({}, this.originalConfig, partialHelperState);
@@ -146,11 +154,19 @@ class URLSync {
       }
     );
 
-    if (this.timer() < this.threshold) {
-      this.urlUtils.replaceState(qs, {getHistoryState: this.getHistoryState});
-    } else {
-      this.urlUtils.pushState(qs, {getHistoryState: this.getHistoryState});
+    if (this.updateOnEveryKeyStroke === true) {
+      if (this.timer() < this.threshold) {
+        this.urlUtils.replaceState(qs, {getHistoryState: this.getHistoryState});
+      } else {
+        this.urlUtils.pushState(qs, {getHistoryState: this.getHistoryState});
+      }
+      return;
     }
+
+    clearTimeout(this.urlUpdateTimeout);
+    this.urlUpdateTimeout = setTimeout(() => {
+      this.urlUtils.pushState(qs, {getHistoryState: this.getHistoryState});
+    }, this.threshold);
   }
 
   // External API's